requirements.txt
mysql.connector
urllib.request 
json
tabulate
os
